floris to pywake turbine translator
Instead of redefining the reference turbines with the same data as floris, it seems more clean to just include a small utility function to read floris input files and return pywake turbine objects. We have reached out to the floris team about the availability of floris turbines using the iea_37 standard ontology, but have not yet found this to be available.